% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkglstring.R
\name{checkglstring}
\alias{checkglstring}
\title{Check GL Strings syntax for locus names appearing in multiple locus fields}
\usage{
checkglstring(glstring)
}
\arguments{
\item{glstring}{A character string containing valid GL string information on alleles at one or more loci.}
}
\value{
NULL if no locus identifiers are found in multiple locus fields, otherwise returns a locus name that appears in multiple locus fields.
}
\description{
This function parses a GL string sufficiently to check if locus names appear in multiple locus fields. This function is used by canonicalize().
}
\details{
Any string containing the GL string delimiters ^, +, /, or | and locus names separated from allele names by * will be parsed and analyzed.
}
