% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambiguate.R
\name{disambiguate}
\alias{disambiguate}
\title{Genotype List Disambiguation Using Human Leukocyte Antigen (HLA) Allele Frequency Data}
\usage{
disambiguate(
  glstrings,
  freqbyacc = NULL,
  deletedalleles = NULL,
  allelehistory = NULL,
  regionstring = "SSA",
  imgtversion = "guess",
  probratio = 0.5,
  dislevel = 0,
  ruledat = gl_match_rules,
  mode = "normal",
  hlaprefix = FALSE,
  log = FALSE,
  allelelog = TRUE,
  verbose = "default",
  dataload = "default"
)
}
\arguments{
\item{glstrings}{A vector of GL strings to be disambiguated.}

\item{freqbyacc}{Frequency-by-accession data frame. If not provided and if \code{loadHLAdata()} can be called succesfully, the parameter defaults to \code{gl_HLA_frequencies_by_accession}, an object that can be created using \code{updateHLAdata()}. Also, a data object in this package, \code{package_HLA_frequencies_by_accession}, contains frequency data. Any data frame with the latter's structure can be passed as this argument and should be useful for providing updated or alternative allele-frequency-by-HLA-accession data.}

\item{deletedalleles}{Deleted alleles data frame. If not provided and if \code{loadHLAdata()} can be called succesfully, the parameter defaults to \code{gl_HLA_deleted_alleles}, an object that can be created using \code{updateHLAdata()}. Also, a data object in this package, \code{package_HLA_deleted_alleles}, contains deleted allele data. Any data frame with this the latter's structure can be passed as this argument, useful for providing updated deleted allele information.}

\item{allelehistory}{IMGT-versioned allele name table. If not provided and if \code{loadHLAdata()} can be called succesfully, the parameter defaults to \code{gl_HLA_allele_history}, an object that can be created using \code{updateHLAdata()}. Also, a data object in this package, \code{package_HLA_allele_history}, contains IMGT allele history data. Any data frame with the latter's structure can be passed as this argument, useful for providing updated IMGT/HLA version data.}

\item{regionstring}{Geographic region designation from the frequency by accession data frame. This is the region for which HLA allele frequency data will be used in genotype frequency calculations. For example, after running \code{updateHLAdata()} successfully, \code{tmp <- disambiguate(region="")} will show the available region designations on the console.}

\item{imgtversion}{Defaults to \code{"guess"}. This parameter is required to specify the version of IMGT/HLA allele namespace as defined in data passed with parameter \code{allelehistory}. For example, after running \code{updateHLAdata()} successfully, \code{tmp <- disambiguate(region="AUS",probratio=1,imgtversion="")} will show the available IMGT versions on the console. If this parameter is set to \code{"guess"}, \code{guessimgtversion} will be passed the GL string(s) to attempt to determine the IMGT/HLA allele version.}

\item{probratio}{Probability ratio <=1 and >= 0 used to maintain ambiguity for possible genotypes at a locus with predicted frequency >= \code{probratio} * maximum genotype frequency calculated for that locus. Note that setting \code{dislevel} to 1 will override \code{probratio} < 1.0. Defaults to 0.5.}

\item{dislevel}{Disambiguation level, where 1 indicates complete disambiguation to single genotypes at each locus, and 0 permits partial disambiguation.}

\item{ruledat}{Disambiguation rules for matching HLA allele names that do not comply perfectly with IMGT version names, defaults to \code{gl_match_rules}, provided by this package.}

\item{mode}{Disambugation mode, where "normal" returns GL strings with alleles as named in input with only "HLA-" stripped off (the prefix can be replaced using the \code{hlaprefix} parameter), "version" returns GL strings with alleles standardized to the names matching the specified version, and "accession" returns GL strings containing accession numbers.}

\item{hlaprefix}{Defaulting to FALSE, if TRUE this parameter will force "HLA-" prefix onto all reported allele names, with the exception of accession numbers.}

\item{log}{Logical to determine if verbose logging of disambiguation calculations is returned with disambiguated GL strings. Defaults to \code{FALSE}.}

\item{allelelog}{Logical to determine if a table of allele name mappings to accession numbers should be returned for each GL string. Defaults to \code{TRUE}.}

\item{verbose}{Default value \code{"default"} provides some console feedback on each disambiguated GL string, \code{"quiet"} silences all messaging by the package with the exception of inconsistency errors, and \code{"verbose"} provides information regarding allele name mapping in addition to default messaging on the console}

\item{dataload}{Controls usage of HLA data tables and by default attempts to load or update and load the tables required for disambiguation. This parameter has the default value \code{"default"}, in which case when versions cannot be determined for \code{allelehistory}, \code{freqbyacc}, or \code{deletedalleles}, the function attempts to load or download and load all the HLA data tables. If the parameter has the value \code{"package"}, then the tables delivered with the package install will be used. If the parameter has the value \code{"argsonly"}, then only when versions can be determined for \code{allelehistory}, \code{freqbyacc}, and \code{deletedalleles} will the function proceed.}
}
\value{
A list with the components:
 \item{data}{A data frame with two variables:}
   \item{$glstrings}{A vector of disambiguated GL strings, one for each input GL string, in the same order as input. Returns NA upon encountering inconsistency processing arguments.}
   \item{$glscores}{A vector of scores, one for each input GL string, in the same order as input.}
 
 \item{alleletable}{A data frame containing five variables:}
   \item{$allele}{The allele for which a match was sought (character).}
   \item{$matched}{The allele string that matched, possibly with special rules applied, or NA (character).}
   \item{$accession}{The accession number assigned the allele, or NA (character).}
   \item{$status}{A factor with possible levels "deleted" (allele was found to have been deleted and replaced), "imgt" (allele was found in the specified IMGT version), "accession-imgt" (allele was identified as an HLA accession number), "rules" (allele was found in the specified IMGT version after applying HLA GL string matching rules), or "unidentified" if an accession number could not be assigned.}

 \item{inconsistency.accession}{TRUE if failures to identify accession numbers occurred, FALSE otherwise. }
 \item{inconsistency.parse}{TRUE if failures to parse input GL strings or process them were encountered, FALSE otherwise.}
 \item{inconsistency.call}{A character string describing an inconsistency. NA if GL strings are returned.}
 \item{log}{Verbose details on GL String disambiguation. NA if argument \code{log} is FALSE.}
}
\description{
This function disambiguates Genotype List (GL) string data based on human leukocyte antigen (HLA) allele frequency data, deleted allele data, geographic region, and IMGT version. Arguements must also be defined to define the behavior of the function with respect to desired output, including complete or partial disambiguation, whether allele names should be standardized to IMGT version data or HLA accession numbers should be output in GL string format.
}
\details{
In this version, "Cw" and "C" are both permitted in glstrings, even if inappropriate to the IMGT version. Use \code{mode="version"} to see the correct allele names for the version. "HLA-" preceding locus designations will be stripped off into order to avoid incorrect allele pairs calculations that could occur in the case that the prefix is used inconsistently within a locus. Haplotype information is not parsed or sorted in this version, so the behavior of disambiguateR for such data is not well-defined. Because disambiguateR calls canonicalize before returning GL strings, the GL strings returned will be reproducible and all input GL strings that describe logically equivalent information will be reported with the same output GL string. If the mode="accession" parameter is set and disambiguation levels are set identically (dislevel,probratio), two GL strings that use different variations of allele names or different IMGT version namespaces and are determined to be semantically equivalent will result in the same output GL string. The glscores are the means of the maximum allele pair frequencies across all loci. When allele pair frequencies cannot be calculated, the maximum allele pair frequency value for a locus is set to zero. The glscores are not affected by complete or partial disambiguation (for example, the \code{dislevel} and \code{probratio} paramaters do not affect the calculation of glscores).
}
\examples{
gls <- c("A*11011/A*11012+A*3303^C*03041/C*0308/C*0309+C*07011/C*07012/C*0706|C*03041/C*0308/C*0309+C*0705|C*0305+C*07011/C*07012/C*0706|C*0310+C*07011/C*07012/C*0706|C*0310+C*0705^B*4002+B*44031/B*44032","HLA-A*0203+HLA-A*2402101/HLA-A*2402101L/HLA-A*24022/HLA-A*2409N/HLA-A*2411N/HLA-A*2420/HLA-A*2426^HLA-B*40011/HLA-B*40012+HLA-B*5401")
# dislevel=0 will perform partial disambiguation. Note extensive log information to stderr.
gls.out <- disambiguate(probratio=0.5,freqbyacc=package_HLA_frequencies_by_accession,deletedalleles=package_HLA_deleted_alleles,allelehistory=package_HLA_allele_history,region="NAF",imgtversion = "1050",dislevel=0,mode="normal",glstrings=gls)
str(gls.out)
gls.out$data
# dislevel=1 will perform complete disambiguation, irrespective of probratio value.
gls.out <- disambiguate(probratio=0.5,freqbyacc=package_HLA_frequencies_by_accession,deletedalleles=package_HLA_deleted_alleles,allelehistory=package_HLA_allele_history,region="NAF",imgtversion = "1050",dislevel=1,mode="normal",glstrings=gls,log=TRUE)
# Top 10 rows of log
cat(paste(strsplit(gls.out$log,"\n")[[1]][1:10],collapse="\n"))
gls.out$data

# not run
# Normally, one will want the latest HLA data tables:
# updateHLAdata()
# The update may take a little time to download.
# Once updated data are updated (or later simply loaded with load_HLA_data), there is no need to specify freqbyacc, deletedalleles, or allelehistory for typical usage:
# gls.out <- disambiguate(probratio=0.5,region="NAF",imgtversion = "1050",dislevel=0,mode="normal",glstrings=gls,log=TRUE)

# For a vector or list of GL Strings, gls, use guessimgtversion to calculate a different imgt version for each GL String, and create a data frame of disambiguated results
# require(dplyr)
# ret <- lapply(1:length(gls), function(X) {print(X);imgt <- guessimgtversion(gls[X]); ret<-disambiguate(gls[X],regionstring="EUR",dislevel=0,mode="normal",probratio=0.8,imgtversion=imgt,log=TRUE,allelelog=FALSE)$data; return(ret)}) \%>\% bind_rows
# ret[1:2,1:4]
# write the verbose log for the second glstring to a file for inspection
# write(ret$log[2],file="HLA_disambiguateR_tmplog.txt")
#
# When there is doubt about the geographic region appropriate for disambiguating GL strings, one can investigate the effect of region on the glscores:
# sapply(unique(gl_HLA_frequencies_by_accession$data$region), function(X) {d <- disambiguate(test.glstrings,region=X,verbose="quiet",dislevel=1);return(sum(d$data$glscores))})

}
