% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchHLAfrequencies.R
\name{fetchHLAfrequencies}
\alias{fetchHLAfrequencies}
\title{fetch HLA frequency data function}
\usage{
fetchHLAfrequencies(url)
}
\arguments{
\item{url}{A url where the HLA frequency text data are found.}
}
\value{
A list object with $data storing the data matrix, $version.date providing a date class variable for the version, and $error with a message or empty string upon successful completion
}
\description{
This function fetches HLA frequency data and is called by updateHLAfrequencies
}
\details{
Called by updateHLAfrequencies
}
\examples{
fret <- fetchHLAfrequencies("https://raw.githubusercontent.com/hughsalamon/disambiguateR/master/inst/HLA_frequencies_by_accession_and_region.txt")
dim(fret$data)
print(fret$version.date)
print(fret$error)

}
