% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchHLAhistory.R
\name{fetchHLAhistory}
\alias{fetchHLAhistory}
\title{fetch HLA history data function}
\usage{
fetchHLAhistory(url)
}
\arguments{
\item{url}{A url where the HLA history text data are found.}
}
\value{
A list object with $data storing the data matrix, $version.date providing a date class variable for the version, and $error with a message or NULL upon successful completion
}
\description{
This function fetches HLA allele name history data and is called by updateHLAhistory
}
\details{
Called by updateHLAhistory
}
\examples{
hret <- fetchHLAhistory("http://ftp.ebi.ac.uk/pub/databases/ipd/imgt/hla/Allelelist_history.txt")
dim(hret$data)
print(hret$version)
print(hret$error)

}
