% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchHLAnomenclature.R
\name{fetchHLAnomenclature}
\alias{fetchHLAnomenclature}
\title{fetch HLA nomenclature data function}
\usage{
fetchHLAnomenclature(url)
}
\arguments{
\item{url}{A url where the HLA nomenclature text data are found.}
}
\value{
A list object with $data storing the data matrix, $version.date providing a date class variable for the version, and $error with a message or empty string upon successful completion
}
\description{
This function fetches HLA nomenclature data and is called by updateHLAnomenclature
}
\details{
Called by updateHLAnomenclature
}
\examples{
fret <- fetchHLAnomenclature("https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/wmda/hla_nom.txt")
dim(fret$data)
print(fret$version.date)
print(fret$error)

}
