% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessimgtversion.R
\name{guessimgtversion}
\alias{guessimgtversion}
\title{IMGT/HLA Version Guessing for Genotype List Disambiguation Using Human Leukocyte Antigen (HLA) Allele Frequency Data}
\usage{
guessimgtversion(glstrings, allelehistory = NULL, alleleload = "default")
}
\arguments{
\item{glstrings}{A vector of glstrings to be disambiguated.}

\item{allelehistory}{Should be the name of an IMGT-versioned allele name table object. If not provided and \code{alleleload} is set to \code{"default"}, as long as the load or download and load is successful, this parameter will be given the value \code{gl_HLA_allele_history} which can be created by this package (see \code{updateHLAdata}). Any data frame with the appropriate structure can be passed as this argument, useful for providing updated IMGT/HLA version data. For example, \code{package_HLA_allele_history} is such a data frame that is provided with this package. If the parameter is not provided and \code{alleleload} is set to \code{"noload"}, this parameter will be given the value \code{package_HLA_allele_history}.}

\item{alleleload}{Can be \code{"default"}, in which case if \code{"allelehistory"} is not defined or does not have a \code{$version} variable, the function will attempt to load current HLA data tables, and offer to download them if they can't be loaded. If set to \code{"noload"}, \code{allelehistory} will be set to  \code{package_HLA_allele_history}. Other values for this parameter will result in an error.}
}
\value{
A string containing the latest version of the IMGT/HLA with the maximal number of allele names matching, NA if no matching alleles found, -1 if there was an error found for the alleleload parameter, -2 if allelehistory could not be loaded.
}
\description{
This function finds the column name in the data specified by the \code{allelehistory} parameter that matches the most allele names found in the GL string data specified by the \code{glstrings} parameter. When more than one column (typically IMGT version) matches equivalently, the column name that sorts highest (most recent IMGT version) is returned.
}
\details{
In this version, Cw and C locus prefixes are treated equivalently.
}
\examples{
gls <- c("A*1101/A*11012+A*3303^B*40:02:01+B*44:03:01/B*44:03:02^Cw*03041/Cw*0308/Cw*0309/Cw*0310+Cw*07011/Cw*07012/Cw*0705/Cw*0706|Cw*0305+Cw*07011/Cw*07012/Cw*0706")
guessimgtversion(glstrings=gls,allelehistory=package_HLA_allele_history)
gls <- c("A*1101/A*11012+A*3303^B*40:02:01+B*44:03:01/B*44:03:02^C*03:04:01:01/C*03:08/C*03:09/C*03:10+C*07:01:01/C*07:01:02/C*07:05/C*07:06|C*03:05+C*07:01:01/C*07:01:02/C*07:06")
guessimgtversion(glstrings=gls,allelehistory=package_HLA_allele_history)

}
