% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHLAdata.R
\name{loadHLAdata}
\alias{loadHLAdata}
\title{HLA data loading function}
\usage{
loadHLAdata()
}
\value{
0 when an load is successful 1 otherwise. See printed messages for clarifications.
}
\description{
This function loads the most current HLA data that has been saved using updateHLAdata()
}
\examples{
loadHLAdata()

}
