% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{package_HLA_deleted_alleles}
\alias{package_HLA_deleted_alleles}
\title{HLA deleted allele records (limited to those where allele names were replaced)}
\format{A list object with three items:
\itemize{
  \item $data: a data frame with 115 observations and 6 variables:
  \itemize{
    \item V1: a factor with 49 levels characterizing HLA loci
    \item V2: a factor with 13718 (largely unused) levels characterizing allele names at the specified loci that were replaced
    \item V3: an integer with date information characterizing the date the allele in $V2 was assigned
    \item V4: an integer with date information characterizing the date the allele in $V5 was assigned (the date of change)
    \item V5: a factor with 112 levels characterizing allele names at the specified loci that replaced the names in $V2
    \item V6: a factor with 6 levels providing the reason for the replacement
  }
  \item $version.date: version date
  \item $error: value "". When using updateHLAdata(), a list object of the same format as package_HLA_deleted_alleles will be created and named gl_HLA_deleted_alleles. In the latter case, errors in creating the object upon attempted download will be recorded in this variable.
}}
\source{
\url{https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/wmda/hla_nom.txt}
}
\description{
A list object containing supporting information and a data frame that documents allele name deprecation and replacements
}
