% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{package_HLA_frequencies_by_accession}
\alias{package_HLA_frequencies_by_accession}
\title{HLA frequencies by accession}
\format{A list object with three items:
\itemize{
  \item $data: a data frame with 115 observations and 8 variables:
  \itemize{
    \item region: a factor with 11 levels characterizing geographic regions, "AUS", "EUR", "NAF", "NAM", "NEA", "OCE", "OTH", "SAM", "SEA", "SSA", or "SWA"
    \item accession: a factor with 2031 levels charactrizing HLA accession numbers
    \item frequency: a numeric giving the frequency of the HLA allele in the geographic region
    \item cwdstatus: a factor with three levels, "Neither reg-CWD nor reg-Rare", "reg-CWD", or "reg-Rare"
    \item cwdp: a factor with three levels, "", "C", "WD"
    \item cwdg: a factor with three levels, "", "C", "WD"
    \item mrwaf.allele: a factor with 840 levels
    \item X390.allele: a factor with 2031 levels
  }
  \item $version.date: version date
  \item $error: value "". When using updateHLAdata(), a list object of the same format as package_HLA_frequencies_by_accession will be created and named gl_HLA_frequencies_by_accession. In the latter case, errors in creating the object upon attempted download will be recorded in this variable.
}}
\source{
\url{http://igdawg.org/pubs/HLA_frequencies_by_accession_and_region.txt}
}
\description{
A list object containing supporting information and a data frame that maps IMGT version 3.9.0 allele names to accession numbers and regional frequency data
}
