% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHLAdata.R
\name{updateHLAdata}
\alias{updateHLAdata}
\title{HLA data updating function}
\usage{
updateHLAdata(force = FALSE, quickcheck = FALSE)
}
\arguments{
\item{force}{FALSE by default. TRUE to force updating independent of loaded object version records.}

\item{quickcheck}{A logical, defaults to FALSE. If TRUE, performs a quick check on HLA history data at \code{url} for the most recent version therein and compares to currently downloaded data. Assumes that the table header of the online data gives the most recent IMGT version string at row 1 column 2. \code{force} takes precedence and if TRUE, quickcheck will be ignored. See \code{updateHLAhistory}.}
}
\value{
NULL. See printed messages for information on update successes or failures.
}
\description{
This function updates HLA data used for disambiguation.
}
\details{
When upgrading the package, if you do not start a clean R session you may have stale objects \code{gl_HLA_nomenclature}, \code{gl_HLA_frequencies_by_accession}, and \code{gl_HLA_history} in the workspace that were reloaded when you started your session. In such cases, \code{updateHLAdata} may fail. You may choose to start a clean R session without restoring the workspace or simply run \code{updateHLAdata(force=TRUE)}.
}
\examples{
updateHLAdata()

}
