% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHLAfrequencies.R
\name{updateHLAfrequencies}
\alias{updateHLAfrequencies}
\title{HLA allele frequencies data updating function}
\usage{
updateHLAfrequencies(
  force = FALSE,
  current = gl_HLA_frequencies_by_accession,
  url =
    "https://raw.githubusercontent.com/hughsalamon/disambiguateR/master/inst/HLA_frequencies_by_accession_and_region.txt"
)
}
\arguments{
\item{force}{Forces updating even if data is older or smaller than the current data.}

\item{current}{A list object containing the current HLA frequency data and version information.}

\item{url}{A url where the HLA frequency text data are found.}
}
\value{
A list object with $status 0 when an update is successful and 1 otherwise. See $msg for clarifications.
}
\description{
This function updates HLA frequency data used for disambiguation.
}
\details{
Calls fetchHLAfrequencies
}
\examples{
loadHLAdata()
uret <- updateHLAfrequencies()
uret$status
uret$msg

}
