% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHLAhistory.R
\name{updateHLAhistory}
\alias{updateHLAhistory}
\title{HLA history data updating function}
\usage{
updateHLAhistory(
  force = FALSE,
  quickcheck = FALSE,
  current = gl_HLA_allele_history,
  url = "http://ftp.ebi.ac.uk/pub/databases/ipd/imgt/hla/Allelelist_history.txt"
)
}
\arguments{
\item{force}{Forces updating even if data is older or smaller than the current data.}

\item{quickcheck}{A logical, defaults to FALSE. If TRUE, performs a quick check on HLA history data at \code{url} for the most recent version therein and compares to currently downloaded data. Assumes that the table header of the online data gives the most recent IMGT version string at row 1 column 2. \code{force} takes precedence and if TRUE, \code{quickcheck} will be ignored.}

\item{current}{A list object containing the current HLA history data and version information.}

\item{url}{A url where the HLA history text data are found.}
}
\value{
A list object with $status 0 when an update is successful and 1 otherwise. See $msg for clarifications.
}
\description{
This function updates HLA history data used for disambiguation.
}
\details{
Calls fetchHLAhistory. \code{quickcheck} is useful if download times for HLA history data are inconvenient.
}
\examples{
#uret <- updateHLAhistory()
#uret$status
#uret$msg

}
