% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHLAnomenclature.R
\name{updateHLAnomenclature}
\alias{updateHLAnomenclature}
\title{HLA nomenclature data updating function}
\usage{
updateHLAnomenclature(
  force = FALSE,
  current = gl_HLA_deleted_alleles,
  url = "https://raw.githubusercontent.com/ANHIG/IMGTHLA/Latest/wmda/hla_nom.txt"
)
}
\arguments{
\item{force}{Forces updating even if data is older or smaller than the current data.}

\item{current}{A list object containing the current HLA nomenclature data and version information.}

\item{url}{A url where the HLA nomenclature text data are found.}
}
\value{
A list object with $status 0 when an update is successful and 1 otherwise. See $msg for clarifications.
}
\description{
This function updates HLA nomenclature data used for disambiguation.
}
\details{
Calls fetchHLAnomenclature
}
\examples{
uret <- updateHLAnomenclature()
uret$status
uret$msg

}
